python-insights/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
apscheduler
aio-pika
python-dotenv
